9.10.21 (Good code checkpoint)
Added file/directory base work

added image interpretation code includes:
 - using PIL to pull images and interpret and manipulate them such as to create the "background" image
 - using cv2 to write temp files which are the concatenation of the original and background adjusted images
 - using cv2 to create .avi file type movie including the side by side comparison

added code that results in accessing external repos (on the users machine, can be any location etc. requires double checking that the folder can be manipulated in this way or if it should not be accessed by the code so as to not be intrusive on users machines).

very experimental build still but a good code checkpoint